+++ solutions/ColumnFormatter/src/webparts/columnFormatter/components/Panels/Tree/NodeProperties.tsx
@@ -236,6 +236,26 @@ export class NodeProperties extends React.Component<INodePropertiesProps, INodeP
 
 		return props;
 	}
+
+	private styleAttributes = [
+		"align-items","background-color","background-image","border","border-bottom","border-bottom-color",
+		"border-bottom-left-radius","border-bottom-right-radius","border-bottom-style","border-bottom-width",
+		"border-collapse","border-color","border-left","border-left-color","border-left-style","border-left-width",
+		"border-radius","border-right","border-right-color","border-right-style","border-right-width","border-spacing",
+		"border-style","border-top","border-top-color","border-top-left-radius","border-top-right-radius","border-top-style",
+		"border-top-width","border-width","bottom","box-align","box-decoration-break","box-direction","box-flex",
+		"box-flex-group","box-lines","box-ordinal-group","box-orient","box-pack","box-shadow","box-sizing","caption-side",
+		"clear","clip","color","column-count","column-fill","column-gap","column-rule","column-rule-color",
+		"column-rule-style","column-rule-width","column-span","column-width","columns","direction","display",
+		"empty-cells","float","font","font-family","font-size","font-size-adjust","font-stretch","font-variant",
+		"font-weight","fill","height","grid-columns","grid-rows","hanging-punctuation","left","letter-spacing",
+		"line-height","margin","margin-bottom","margin-left","margin-right","margin-top","max-height","max-width",
+		"min-height","min-width","opacity","outline","outline-color","outline-style","outline-width","overflow",
+		"overflow-x","overflow-y","padding","padding-bottom","padding-left","padding-right","padding-top","position",
+		"punctuation-trim","right","rotation","rotation-point","table-layout","text-align","text-align-last","text-decoration",
+		"text-indent","text-justify","text-outline","text-shadow","text-transform","text-wrap","top","unicode-bidi",
+		"vertical-align","visibility","white-space","width","word-break","word-spacing","word-wrap","z-index",
+	];
 	
 	private buildProp(propertyName:string, addressPrefix:string, node:any, elmType:string, formatType:formatterType, relevantPattern?:RegExp): INodeProperty {
 		const isCurrent:boolean = (typeof node !== "undefined" && node.hasOwnProperty(propertyName));
@@ -277,6 +297,49 @@ export class NodeProperties extends React.Component<INodePropertiesProps, INodeP
 			case "elmType":
 			case "target":
 			case "action":
+			case "align-items":
+			case "border-bottom-style":
+			case "border-left-style":
+			case "border-right-style":
+			case "border-style":
+			case "border-top-style":
+			case "box-align":
+			case "box-decoration-break":
+			case "box-direction":
+			case "box-lines":
+			case "box-orient":
+			case "box-pack":
+			case "box-sizing":
+			case "caption-side":
+			case "clear":
+			case "column-fill":
+			case "column-rule-style":
+			case "column-span":
+			case "direction":
+			case "display":
+			case "empty-cells":
+			case "float":
+			case "font-stretch":
+			case "font-variant":
+			case "hanging-punctuation":
+			case "outline-style":
+			case "overflow":
+			case "overflow-x":
+			case "overflow-y":
+			case "position":
+			case "punctuation-trim":
+			case "table-layout":
+			case "text-align":
+			case "text-align-last":
+			case "text-decoration":
+			case "text-justify":
+			case "text-transform":
+			case "text-wrap":
+			case "unicode-bidi":
+			case "visibility":
+			case "white-space":
+			case "word-break":
+			case "word-wrap":
 				return NodePropType.dropdown;
 			case "class":
 			case "role":
@@ -287,6 +350,86 @@ export class NodeProperties extends React.Component<INodePropertiesProps, INodeP
 			case "hideSelection":
 			case "hideListHeader":
 				return NodePropType.toggle;
+			case "background-color":
+			case "border-bottom-color":
+			case "border-color":
+			case "border-left-color":
+			case "border-right-color":
+			case "border-top-color":
+			case "color":
+			case "column-rule-color":
+			case "fill":
+			case "outline-color":
+				return NodePropType.color;
+			case "background-image":
+			case "border":
+			case "border-bottom":
+			case "border-bottom-left-radius":
+			case "border-bottom-right-radius":
+			case "border-bottom-width":
+			case "border-left":
+			case "border-left-width":
+			case "border-radius":
+			case "border-right":
+			case "border-right-width":
+			case "border-spacing":
+			case "border-top":
+			case "border-top-left-radius":
+			case "border-top-right-radius":
+			case "border-top-width":
+			case "border-width":
+			case "bottom":
+			case "box-flex":
+			case "box-flex-group":
+			case "box-ordinal-group":
+			case "box-shadow":
+			case "clip":
+			case "column-count":
+			case "column-gap":
+			case "column-rule":
+			case "column-rule-width":
+			case "column-width":
+			case "columns":
+			case "font":
+			case "font-family":
+			case "font-size":
+			case "font-size-adjust":
+			case "font-weight":
+			case "height":
+			case "grid-columns":
+			case "grid-rows":
+			case "left":
+			case "letter-spacing":
+			case "line-height":
+			case "margin":
+			case "margin-bottom":
+			case "margin-left":
+			case "margin-right":
+			case "margin-top":
+			case "max-height":
+			case "max-width":
+			case "min-height":
+			case "min-width":
+			case "opacity":
+			case "outline":
+			case "outline-width":
+			case "padding":
+			case "padding-bottom":
+			case "padding-left":
+			case "padding-right":
+			case "padding-top":
+			case "right":
+			case "rotation":
+			case "rotation-point":
+			case "text-indent":
+			case "text-outline":
+			case "text-shadow":
+			case "top":
+			case "vertical-align":
+			case "width":
+			case "word-spacing":
+			case "z-index":
+				return NodePropType.combobox;
 			default:
 				return NodePropType.text;
 		}
@@ -589,9 +732,308 @@ export class NodeProperties extends React.Component<INodePropertiesProps, INodeP
 				];*/
 				//console.log(classes.join('","'));
 				return this.stringsToOptions(classes);
+
+			
+			//Style attributes	
+			case "align-items":
+				return this.styleStringsToOptions([
+					"stretch","center","flex-start","flex-end","baseline",
+				]);
+			case "background-color":
+			case "border-bottom-color":
+			case "border-color":
+			case "border-left-color":
+			case "border-right-color":
+			case "border-top-color":
+			case "color":
+			case "column-rule-color":
+			case "fill":
+				return this.styleStringsToOptions(this.namedColors);
+			case "background-image":
+			case "column-count":
+			case "font-size-adjust":
+			case "grid-columns":
+			case "grid-rows":
+			case "max-height":
+			case "max-width":
+			case "text-shadow":
+				return this.styleStringsToOptions([
+					"none"
+				]);
+			case "border": //styleStandard
+			case "border-bottom":
+			case "border-bottom-left-radius":
+			case "border-bottom-right-radius":
+			case "border-left":
+			case "border-radius":
+			case "border-right":
+			case "border-spacing":
+			case "border-top":
+			case "border-top-left-radius":
+			case "border-top-right-radius":
+			case "box-flex":
+			case "box-flex-group":
+			case "box-ordinal-group":
+			case "column-rule":
+			case "font-family":
+			case "min-height":
+			case "min-width":
+			case "opacity":
+			case "outline":
+			case "padding":
+			case "padding-bottom":
+			case "padding-left":
+			case "padding-right":
+			case "padding-top":
+			case "rotation":
+			case "rotation-point":
+			case "text-indent":
+			case "text-outline":
+				return this.stringsToOptions(this.styleBasic);
+			case "border-bottom-style":
+			case "border-left-style":
+			case "border-right-style":
+			case "border-style":
+			case "border-top-style":
+			case "column-rule-style":
+			case "outline-style":
+				return this.styleStringsToOptions([
+					"none","hidden","dotted","dashed","solid","double","groove","ridge","inset","outset",
+				]);
+			case "border-bottom-width":
+			case "border-left-width":
+			case "border-right-width":
+			case "border-top-width":
+			case "border-width":
+			case "column-rule-width":
+			case "outline-width":
+				return this.styleStringsToOptions([
+					"thin","medium","thick",
+				]);
+			case "border-collapse":
+				return this.styleStringsToOptions([
+					"separate","collapse",
+				]);
+			case "bottom": //styleStandardAuto
+			case "clip":
+			case "column-width":
+			case "columns":
+			case "height":
+			case "left":
+			case "margin":
+			case "margin-bottom":
+			case "margin-left":
+			case "margin-right":
+			case "margin-top":
+			case "right":
+			case "top":
+			case "width":
+			case "z-index":
+				return this.styleStringsToOptions([],true);
+			case "box-align":
+				return this.styleStringsToOptions([
+					"start","center","end","baseline","stretch",
+				]);
+			case "box-decoration-break":
+				return this.styleStringsToOptions([
+					"slice","clone",
+				]);
+			case "box-direction":
+				return this.styleStringsToOptions([
+					"normal","reverse",
+				]);
+			case "box-lines":
+				return this.styleStringsToOptions([
+					"single","multiple",
+				]);
+			case "box-orient":
+				return this.styleStringsToOptions([
+					"horizontal","vertical","inline-axis","block-axis",
+				]);
+			case "box-pack":
+				return this.styleStringsToOptions([
+					"start","center","end","justify",
+				]);
+			case "box-shadow":
+				return this.styleStringsToOptions([
+					"none","inset",
+				]);
+			case "box-sizing":
+				return this.styleStringsToOptions([
+					"content-box","border-box",
+				]);
+			case "caption-side":
+				return this.styleStringsToOptions([
+					"top","bottom",
+				]);
+			case "clear":
+				return this.styleStringsToOptions([
+					"none","left","right","both",
+				]);
+			case "column-fill":
+				return this.styleStringsToOptions([
+					"balance",
+				],true);
+			case "column-gap":
+				return this.styleStringsToOptions([
+					"normal",
+				]);
+			case "column-span":
+				return this.styleStringsToOptions([
+					"none","all",
+				]);
+			case "direction":
+				return this.styleStringsToOptions([
+					"ltr","rtl",
+				]);
+			case "display":
+				return this.styleStringsToOptions([
+					"inline","block","flex","inline-block","inline-flex","inline-table","list-item",
+					"run-in","table","table-caption","table-column-group","table-header-group",
+					"table-row-group","table-cell","table-column","table-row","none",
+				]);
+			case "empty-cells":
+				return this.styleStringsToOptions([
+					"show","hide",
+				]);
+			case "float":
+				return this.styleStringsToOptions([
+					"none","left","right",
+				]);
+			case "font":
+				return this.styleStringsToOptions([
+					"caption","icon","menu","message-box","small-caption","status-bar",
+				]);
+			case "font-size":
+				return this.styleStringsToOptions([
+					"medium","xx-small","x-small","small","large","x-large","xx-large","smaller","larger",
+				]);
+			case "font-stretch":
+				return this.styleStringsToOptions([
+					"ultra-condensed","extra-condensed","condensed","semi-condensed",
+					"normal","semi-expanded","expanded","extra-expanded","ultra-expanded",
+				]);
+			case "font-variant":
+				return this.styleStringsToOptions([
+					"normal","small-caps",
+				]);
+			case "font-weight":
+				return this.styleStringsToOptions([
+					"normal","bold","bolder","lighter",
+				]);
+			case "hanging-puncuation":
+				return this.styleStringsToOptions([
+					"none","first","last","allow-end","force-end",
+				]);
+			case "letter-spacing":
+			case "line-height":
+			case "word-spacing":
+				return this.styleStringsToOptions([
+					"normal",
+				]);
+			case "outline-color":
+				return this.styleStringsToOptions([
+					"invert",
+				].concat(this.namedColors));
+			case "overflow":
+			case "overflow-x":
+			case "overflow-y":
+				return this.styleStringsToOptions([
+					"visible","hidden","scroll",
+				],true);
+			case "position":
+				return this.styleStringsToOptions([
+					"static","absolute","fixed","relative","sticky",
+				]);
+			case "punctuation-trim":
+				return this.styleStringsToOptions([
+					"none","start","end","allow-end","adjacent",
+				]);
+			case "table-layout":
+				return this.styleStringsToOptions([
+					"auto","fixed",
+				]);
+			case "text-align":
+				return this.styleStringsToOptions([
+					"left","right","center","justify",
+				]);
+			case "text-align-last":
+				return this.styleStringsToOptions([
+					"auto","left","right","center","justify","start","end",
+				]);
+			case "text-decoration":
+				return this.styleStringsToOptions([
+					"none","underline","overline","line-through",
+				]);
+			case "text-justify":
+				return this.styleStringsToOptions([
+					"auto","none","inter-word","inter-ideograph","inter-cluster","distribute","kashida","trim",
+				]);
+			case "text-transform":
+				return this.styleStringsToOptions([
+					"none","capitalize","uppercase","lowercase",
+				]);
+			case "text-wrap":
+				return this.styleStringsToOptions([
+					"normal","none","unrestricted","suppress",
+				]);
+			case "unicode-bidi":
+				return this.styleStringsToOptions([
+					"normal","embed","bidi-override",
+				]);
+			case "vertical-align":
+				return this.styleStringsToOptions([
+					"baseline","sub","super","top","text-top","middle","bottom","text-bottom",
+				]);
+			case "visibility":
+				return this.styleStringsToOptions([
+					"visible","hidden","collapse",
+				]);
+			case "white-space":
+				return this.styleStringsToOptions([
+					"normal","nowrap","pre","pre-line","pre-wrap",
+				]);
+			case "word-break":
+				return this.styleStringsToOptions([
+					"normal","break-all","keep-all",
+				]);
+			case "word-wrap":
+				return this.styleStringsToOptions([
+					"normal","break-word",
+				]);
 			default:
 				return undefined;
 		}
+	}
+
+	private styleBasic = [
+		"initial","inherit","unset"
+	];
+
+	private namedColors = [
+		"transparent","currentColor",
+		"aqua","black","blue","fuchsia","gray","green","lime","maroon","navy","olive","purple","red","silver","teal","white","yellow",
+		"aliceblue","antiquewhite","aquamarine","azure","beige","bisque","blanchedalmond","blueviolet","brown","burlywood","cadetblue",
+		"chartreuse","chocolate","coral","cornflowerblue","cornsilk","crimson","cyan","darkblue","darkcyan","darkgoldenrod","darkgray",
+		"darkgreen","darkkhaki","darkmagenta","darkolivegreen","darkorange","darkorchid","darkred","darksalmon","darkseagreen",
+		"darkslateblue","darkslategray","darkturquoise","darkviolet","deeppink","deepskyblue","dimgray","dodgerblue","firebrick",
+		"floralwhite","forestgreen","gainsboro","ghostwhite","gold","goldenrod","greenyellow","honeydew","hotpink","indianred","indigo",
+		"ivory","khaki","lavender","lavenderblush","lawngreen","lemonchiffon","lightblue","lightcoral","lightcyan","lightgoldenrodyellow",
+		"lightgray","lightgreen","lightpink","lightsalmon","lightseagreen","lightskyblue","lightslategray","lightsteelblue","lightyellow",
+		"limegreen","linen","magenta","mediumaquamarine","mediumblue","mediumorchid","mediumpurple","mediumseagreen","mediumslateblue",
+		"mediumspringgreen","mediumturquoise","mediumvioletred","midnightblue","mintcream","mistyrose","moccasin","navajowhite","oldlace",
+		"olivedrab","orange","orangered","orchid","palegoldenrod","palegreen","paleturquoise","palevioletred","papayawhip","peachpuff",
+		"peru","pink","plum","powderblue","rosybrown","royalblue","saddlebrown","salmon","sandybrown","seagreen","seashell","sienna",
+		"skyblue","slateblue","slategray","snow","springgreen","steelblue","tan","thistle","tomato","turquoise","violet","wheat",
+		"whitesmoke","yellowgreen","rebeccapurple"
+	];
+
+	private styleStringsToOptions(values: Array<string>, auto:boolean = false): Array<IDropdownOption> {
+		if(auto) {
+			values.push("auto");
+		}
+		values.push(...this.styleBasic);
+		return this.stringsToOptions(values);
 	}
 
 	private stringsToOptions(values: Array<string>):  Array<IDropdownOption> {